requirements.txt
django~=3.1.6
djangorestframework~=3.11.0
black~=20.8b1
